Prevent the creation of an empty replace_mmv.xlsx
Now, the file is only created when needed (if 'MMV' is found in the building combinations) and otherwise "None" is passed. The MMV archetype precalculation stays the same (again, only if 'MMV' is passed in the building combinations), i.e., the MMV idf file (e.g. SFH_auto-MMV.idf) is created and then subsequent runs use this pre-created idf file.