IDF Converter Updates
Hi Kamila, I updated the IDF converter and its documentation. I took away the surrogate element additions completely. I realized the BuildME code is changed significantly, so adding frames will be cumbersome, and is not critical at all in this stage I believe. Same with the automatic DHW calculations and etc.  As a final thing we can take a look the result collection one last time together and decide whether we want to use the html files or the epsout file for collecting the final values.